Changed the data type to double Added test case for summary data added code comments

+++ src/main/java/com/expense/model/Member.java
@@ -12,12 +12,12 @@ public class Member implements Comparable<Member>{
 	private String memberId;
 	private String memberName;
 	private String email;
-	private int balance;
+	private double balance;
 	
-	public int getBalance() {
+	public double getBalance() {
 		return balance;
 	}
-	public void setBalance(int balance) {
+	public void setBalance(double balance) {
 		this.balance = balance;
 	}
 	public String getMemberId() {
